requirements.txt
# Make sure clams-python version is explicitly specified, at least the lower bound
clams-python==1.2.1
mmif-python[cv]
python-doctr[torch]==0.8.1
torch>=2.*
numpy
Pillow
lapps
